+++ tests/profiles/universal_test.py
@@ -779,3 +779,23 @@ def test_check_unreachable_glyphs():
                   'ampersand.sc',
                   'I.uc']:
         assert glyph in message
+
+
+def test_check_contour_count(montserrat_ttFonts):
+    """Check glyphs contain the recommended contour count"""
+    check = CheckTester(universal_profile,
+                        "com.google.fonts/check/contour_count")
+    # TODO: FAIL, "lacks-cmap"
+
+    for ttFont in montserrat_ttFonts:
+        # Montserrat which was used to assemble the glyph data,
+        # so, it should be good, except for that softhyphen...
+        # TODO: how can we test PASS then?
+        assert_results_contain(check(ttFont),
+                               WARN, 'softhyphen')
+
+    # Lets swap the glyf 'a' (2 contours) with glyf 'c' (1 contour)
+    for ttFont in montserrat_ttFonts:
+        ttFont['glyf']['a'] = ttFont['glyf']['c']
+        assert_results_contain(check(ttFont),
+                               WARN, 'contour-count')